+++ assets/js/app/certificates/certificates-controller.js
@@ -25,12 +25,72 @@
                     ariaLabelledBy: 'modal-title',
                     ariaDescribedBy: 'modal-body',
                     templateUrl: 'js/app/certificates/add-certificates-modal.html',
-                    controller: function($scope,$uibModalInstance,_certificate){
+                    controller: function($scope,$uibModal,$uibModalInstance,SnisModel,DialogService,_certificate){
                         $scope.update = _certificate
                         $scope.data = _certificate || {}
                         $scope.close = function() {
                             return $uibModalInstance.dismiss()
                         }
+
+                        $scope.deleteSNI = function(sni) {
+                            DialogService.prompt(
+                                "Confirm", "Really want to delete the selected item?",
+                                ['No don\'t', 'Yes! delete it'],
+                                function accept() {
+                                    SnisModel.delete(sni)
+                                        .then(function (res) {
+
+                                            $scope.data.snis.splice($scope.data.snis.indexOf(sni), 1);
+                                        },function(err){
+                                            $log.error("ListConfigService : Model delete failed => ",err)
+                                        });
+                                }, function decline() {
+                                })
+                        }
+
+
+                        $scope.openAddSniModal = function() {
+                            var modalInstance = $uibModal.open({
+                                animation: true,
+                                ariaLabelledBy: 'modal-title',
+                                ariaDescribedBy: 'modal-body',
+                                templateUrl: 'js/app/certificates/add-sni-modal.html',
+                                size: 'sm',
+                                controller: function($scope,$uibModalInstance,SnisModel,DialogService,_certId){
+
+                                    $scope.close = function() {
+                                        return $uibModalInstance.dismiss();
+                                    };
+
+                                    $scope.submit = function() {
+                                        SnisModel.create({
+                                            name : $scope.sni,
+                                            ssl_certificate_id : _certId
+                                        }).then(function(created){
+                                            $uibModalInstance.close({
+                                                data : created.data.name
+                                            });
+                                        }).catch(function (err) {
+                                           console.error("Failed to create SNI =>", err);
+                                            SnisModel.handleError($scope,err);
+                                        });
+                                    };
+
+                                },
+                                resolve : {
+                                    _certId : function() {
+                                        return _certificate.id;
+                                    }
+                                }
+
+                            });
+
+                            modalInstance.result.then(function (data) {
+                                if(data && data.data) $scope.data.snis.push(data.data);
+                            }, function (data) {
+                            });
+                        }
+
 
 
                         $scope.submitCerts = function() {
@@ -65,13 +125,10 @@
                                 var progressPercentage = parseInt(100.0 * evt.loaded / evt.total);
                                 console.log('progress: ' + progressPercentage + '% ' + evt.config.data.file.name);
                             });
+
                         }
 
-                        modalInstance.result.then(function () {
 
-                        }, function (data) {
-                            if(data && data.data) _fetchData()
-                        });
 
                         function handleErrors(err) {
                             $scope.errors = {}
@@ -104,6 +161,12 @@
                         }
                     }
                     //size: 'lg',
+                });
+
+                modalInstance.result.then(function () {
+
+                }, function (data) {
+                    if(data && data.data) _fetchData()
                 });
             }